Refactor preparing for new orionldGetEntity
