+++ src/lib/orionld/mongoc/CMakeLists.txt
@@ -68,6 +68,7 @@ SET (SOURCES
     mongocRegistrationDelete.cpp
     mongocRegistrationReplace.cpp
     mongocRegistrationsIter.cpp
+    mongocAuxAttributesFilter.cpp
 )
 
 # Include directories
