+++ src/lib/orionld/serviceRoutines/orionldPutEntity.cpp
@@ -265,7 +265,7 @@ bool orionldPutEntity(void)
   //
   // Get the entity from the database
   //
-  KjNode* oldDbEntityP = mongocEntityLookup(entityId);
+  KjNode* oldDbEntityP = mongocEntityLookup(entityId, NULL, NULL);
   if (oldDbEntityP == NULL)
   {
     orionldError(OrionldResourceNotFound, "Entity does not exist", entityId, 404);
